# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Postcard API",
    "description": "デジタル絵葉書のリレーアプリケーションAPI",
    "version": "1.0.0"
  },
  "paths": {
    "/api/users": {
      "post": {
        "tags": ["users"],
        "summary": "ユーザープロフィール作成",
        "description": "Amazon Cognitoでの新規登録後、ユーザーのプロフィール情報を初めて作成する際に利用します。",
        "operationId": "create_user_profile_api_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "リクエストボディが不正な場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証トークンがない、または無効な場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "すでにプロフィールが存在する場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "tags": ["users"],
        "summary": "自身のユーザー情報取得",
        "description": "認証済みユーザー自身の情報を取得します。",
        "operationId": "get_my_profile_api_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "認証トークンがない、または無効な場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "ユーザープロフィールが見つからない場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": ["users"],
        "summary": "自身のユーザー情報更新",
        "description": "認証済みユーザー自身の情報を更新します。",
        "operationId": "update_my_profile_api_users_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "リクエストボディが不正な場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証トークンがない、または無効な場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": ["users"],
        "summary": "自身のユーザー削除",
        "description": "認証済みユーザー自身を削除します。**注意: この操作は取り消せません。**",
        "operationId": "delete_my_profile_api_users_me_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証トークンがない、または無効な場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": ["users"],
        "summary": "他のユーザー情報取得",
        "description": "指定した user_id のユーザー情報を取得します。他のユーザーのプロフィール閲覧に利用します。",
        "operationId": "get_user_profile_api_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicProfile"
                }
              }
            }
          },
          "401": {
            "description": "認証トークンがない、または無効な場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "指定したユーザーIDが見つからない場合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "type": "object",
        "required": ["detail"],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserCreateRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "profile_image_url": {
            "type": "string",
            "title": "Profile Image Url"
          }
        },
        "type": "object",
        "required": ["username", "profile_image_url"],
        "title": "UserCreateRequest"
      },
      "UserCreateResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": ["user_id", "message"],
        "title": "UserCreateResponse"
      },
      "UserDeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "UserDeleteResponse"
      },
      "UserProfile": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "profile_image_url": {
            "type": "string",
            "title": "Profile Image Url"
          }
        },
        "type": "object",
        "required": ["user_id", "username", "email", "profile_image_url"],
        "title": "UserProfile"
      },
      "UserPublicProfile": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "profile_image_url": {
            "type": "string",
            "title": "Profile Image Url"
          }
        },
        "type": "object",
        "required": ["user_id", "username", "profile_image_url"],
        "title": "UserPublicProfile"
      },
      "UserUpdateRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "profile_image_url": {
            "type": "string",
            "title": "Profile Image Url"
          }
        },
        "type": "object",
        "required": ["username", "profile_image_url"],
        "title": "UserUpdateRequest"
      },
      "UserUpdateResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": ["message", "user_id"],
        "title": "UserUpdateResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}